requirements.txt
fastapi
mysql-connector-python
uvicorn
python-dotenv